Conflicts:
	Adventure/bin/computc/GameData.class
	Adventure/bin/computc/entities/Enemy.class
	Adventure/bin/computc/entities/Entity.class
	Adventure/bin/computc/entities/Hero.class
	Adventure/bin/computc/states/MainGameState.class
	Adventure/bin/computc/worlds/Dungeon.class
	Adventure/bin/computc/worlds/FiveRoomDungeon.class
	Adventure/bin/computc/worlds/OneRoomDungeon.class
	Adventure/bin/computc/worlds/PredesignedDungeon.class
	Adventure/bin/computc/worlds/RandomRoguelikeDungeon.class
	Adventure/bin/computc/worlds/RandomZeldaesqueDungeon.class
	Adventure/bin/computc/worlds/Room.class
	Adventure/res/rooms/empty.room.tmx
	Adventure/res/rooms/threelines.room.tmx
	Adventure/res/rooms/twolines.room.tmx
	Adventure/src/computc/GameData.java
	Adventure/src/computc/entities/Enemy.java
	Adventure/src/computc/worlds/FiveRoomDungeon.java
	Adventure/src/computc/worlds/OneRoomDungeon.java
	Adventure/src/computc/worlds/PredesignedDungeon.java
	Adventure/src/computc/worlds/RandomRoguelikeDungeon.java
	Adventure/src/computc/worlds/RandomZeldaesqueDungeon.java
	Adventure/src/computc/worlds/Room.java
	Adventure/src/computc/worlds/dungeons/Dungeon.java

+++ Adventure/src/computc/entities/BigThug.java
@@ -3,12 +3,12 @@ package computc.entities;
 import org.newdawn.slick.Image;
 import org.newdawn.slick.SlickException;
 
-import computc.worlds.Dungeon;
-import computc.worlds.Room;
+import computc.worlds.dungeons.Dungeon;
+import computc.worlds.rooms.Room;
 
 public class BigThug extends Thug
 {
-	public BigThug(Dungeon dungeon, Room room, float x, float y) throws SlickException
+	public BigThug(Dungeon dungeon, Room room, int x, int y) throws SlickException
 	{
 		super(dungeon, room, x, y);
